Factor out common code.
