+++ local-modules/bayou-mocha/ClientTests.js
@@ -2,10 +2,9 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import fs from 'fs';
-import path from 'path';
-
 import { Dirs } from 'server-env';
+
+import Utils from './Utils';
 
 /**
  * Driver for the Mocha framework, for client tests.
@@ -19,15 +18,6 @@ export default class ClientTests {
    * @returns {array<string>} Array of module names.
    */
   static moduleNames() {
-    // TODO: This is largely duplicative of similar code in `ServerTests`. It
-    // should be refactored.
-    const packageData = fs.readFileSync(path.resolve(Dirs.CLIENT_DIR, 'package.json'));
-    const packageParsed = JSON.parse(packageData);
-    const dependencies = packageParsed['dependencies'];
-    const modules = Object.keys(dependencies);
-
-    return modules.filter((module) => {
-      return dependencies[module].indexOf('local-modules/') >= 0;
-    });
+    return Utils.localModulesIn(Dirs.CLIENT_DIR);
   }
 }
